Adding a sample file with assertions
