+++ package.json
@@ -7,7 +7,7 @@
     "test": "tests"
   },
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha tests --timeout 10000"
   },
   "repository": {
     "type": "git",
@@ -20,7 +20,9 @@
   },
   "homepage": "https://github.com/knod/docassemble-juvenilesealing#readme",
   "dependencies": {
+    "chai": "^4.2.0",
     "dotenv": "^8.2.0",
+    "mocha": "^7.2.0",
     "puppeteer": "^3.1.0"
   }
 }
